Moved parsing logic to two classes
An ASM Parser handles an input file line by line, and passes each line
to an Opcode Parser that returns a list of instructions for each line.

Main file -> ASM Parser -> Opcode Parser

If a level encounters a syntax or logic error, an exception is thrown,
and caught by the level above.

Updated README to reflect changes in this commit